requirements.txt
fastapi
uvicorn
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
groq
python-dotenv
